pyproject.toml
[tool.poetry]
name = "concourse-gitea-pr"
version = "0.1.2"
description = ""
authors = ["Nuno Goncalves <nuno@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.27.1"
python-dateutil = "^2.8.2"
isort = "^5.10.1"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
pytest = "^6.2.5"
flake8 = "^4.0.1"
mypy = "^0.931"
types-python-dateutil = "^2.8.9"
types-requests = "^2.27.8"
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
check = "concourse_gitea_pr.check:main"
in = "concourse_gitea_pr.in_:main"
out = "concourse_gitea_pr.out:main"

[tool.isort]
profile = "black"
